Split core module in mutiple files
