+++ src/builtins/core/std.h
@@ -1,0 +1,14 @@
+#ifndef STD_H
+#define STD_H
+
+#include "jstar.h"
+
+JSR_NATIVE(jsr_ascii);
+JSR_NATIVE(jsr_char);
+JSR_NATIVE(jsr_eval);
+JSR_NATIVE(jsr_garbageCollect);
+JSR_NATIVE(jsr_int);
+JSR_NATIVE(jsr_print);
+JSR_NATIVE(jsr_type);
+
+#endif
