+++ src/vm.c
@@ -5,6 +5,7 @@
 
 #include "builtins/builtins.h"
 #include "builtins/core/core.h"
+#include "builtins/core/excs.h"
 #include "code.h"
 #include "disassemble.h"
 #include "gc.h"
